crypto_ops.py
# crypto_ops.py
import numpy as np
from concrete.numpy import compile_numpy_function
import pickle

def generate_keys():
    """Generate keys for homomorphic encryption using Concrete-Numpy."""
    # Define a simple function to be compiled
    def quantized_func(x):
        return x + 42
    
    # Compile the function with Concrete-Numpy
    compiler = compile_numpy_function(
        quantized_func,
        {"x": "encrypted"}
    )
    
    # Generate keys
    public_key = np.random.bytes(32)  # Simplified for demo
    private_key = compiler.client_setup()
    evaluation_key = compiler.server_setup()
    
    return public_key, private_key, evaluation_key

def encrypt_data(data: np.ndarray, public_key: bytes) -> np.ndarray:
    """Encrypt data using Concrete-Numpy."""
    # Create a simple encryption function
    def encryption_func(x):
        return x
    
    compiler = compile_numpy_function(
        encryption_func,
        {"x": "encrypted"}
    )
    
    # Encrypt the data
    encrypted = compiler.encrypt(data)
    return encrypted

def decrypt_data(encrypted_data: np.ndarray, private_key: bytes) -> np.ndarray:
    """Decrypt data using Concrete-Numpy."""
    # Create a simple decryption function
    def decryption_func(x):
        return x
    
    compiler = compile_numpy_function(
        decryption_func,
        {"x": "encrypted"}
    )
    
    # Decrypt the data
    decrypted = compiler.decrypt(encrypted_data)
    return decrypted

def evaluate_data(encrypted_data: np.ndarray, evaluation_key: bytes) -> np.ndarray:
    """Perform homomorphic operations on encrypted data."""
    # Define evaluation function
    def evaluation_func(x):
        return x + 1  # Simple increment operation
    
    compiler = compile_numpy_function(
        evaluation_func,
        {"x": "encrypted"}
    )
    
    # Evaluate the encrypted data
    evaluated = compiler.run(encrypted_data, evaluation_key)
    return evaluated

def save_key(key: bytes, filename: str):
    """Save key to binary file."""
    with open(filename, 'wb') as f:
        pickle.dump(key, f)

def load_key(filename: str) -> bytes:
    """Load key from binary file."""
    with open(filename, 'rb') as f:
        return pickle.load(f)