requirements.txt
# requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.24.3
concrete-numpy==0.9.0